+++ migration/data/Version20230614120802.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\CountryVatAdministration\Migrations;
+namespace OxidEsales\CountryVat\Migrations;
 
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;